+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSStonebrickStrongholdStructure.java
@@ -30,9 +30,31 @@ public class RSStonebrickStrongholdStructure extends StrongholdFeature {
         super(DefaultFeatureConfig.CODEC);
     }
 
+    // Thickness of rings:      1,536  (96 chunks)
+    // Distance between rings:  1,536  (96 chunks)
+    // Distance to first ring:  1,280  (80 chunks)
+    // Vanilla has 8 rings
+
     @Override
-    protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeSource biomeSource, long seed, ChunkRandom chunkRandom, int x, int z, Biome biome, ChunkPos chunkPos, DefaultFeatureConfig featureConfig) {
-        return (x * x) + (z * z) > 10000;
+    protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeSource biomeSource, long seed, ChunkRandom chunkRandom, int xChunk, int zChunk, Biome biome, ChunkPos chunkPos, DefaultFeatureConfig featureConfig) {
+        int ringThickness = 96;
+        int distanceToFirstRing = 80;
+
+        int chunkDistance = (int) Math.sqrt((xChunk * xChunk) + (zChunk * zChunk));
+
+        // Offset the distance so that the first ring is closer to spawn
+        int shiftedChunkDistance = Math.max(chunkDistance + (ringThickness - distanceToFirstRing), 0);
+
+        // Determine which ring we are in.
+        // non-stronghold rings are even number ringSection
+        // stronghold rings are odd number ringSection.
+        int ringSection = shiftedChunkDistance / ringThickness;
+
+        // Would mimic vanilla's 8 ring result
+        // if(ringSection > 16) return false;
+
+        // Only spawn strongholds on odd number sections
+        return ringSection % 2 == 1;
     }
 
     @Override